ssss

+++ src/routes/usuario.js
@@ -58,11 +58,16 @@ router.put("/usuario", verificarAutenticacao,  async (req, res) => {
   }
 });
 
-router.delete("/usuario/:id",   async (req, res) => {
+router.delete("/usuario/:id", verificarAutenticacao, async (req, res) => {
   console.log("Rota DELETE /usuario solicitada");
+  // req.userId
   try {
-    await deleteUsuario(req.params.id);
+    if (req.userId == req.params.id){
+      await deleteUsuario(req.params.id);
     res.status(200).json({ message: "Usuário excluido com sucesso!" });
+    } else {
+      res.status(401).json({ message: "Não foi possível excluir este usuário" });
+    }
   } catch (error) {
     res.status(error.status || 500).json({ message: error.message || "Erro!" });
   }
